Setup controllers

+++ database/migrations/2019_07_08_184422_create_comments_table.php
@@ -10,8 +10,8 @@ class CreateCommentsTable extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('id_campaign_from');
-            $table->bigIncrements('id_campaign_to');
+            $table->unsignedInteger('id_campaign_from');
+            $table->unsignedInteger('id_campaign_to');
             $table->text('text');
             $table->timestamps();
         });
